Animation/Sound
added new animations, and sounds that vary randomly in pitch to the clicker.

The animations play 3 different animations when clicked in succession, resulting in a combo animation.

It plays a sound effect upon being clicked that is varied randomly in pitch (so the poor player doesnt have to listen to the same sound EVERY single time they click)